pyproject.toml
[project]
name = "medical-management-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mysql-connector-python>=9.3.0",
    "tkcalendar>=1.6.1",
    "pillow (>=11.2.1,<12.0.0)",
]
